feat(schemas): define tenant related type and guard in schemas/models
