+++ packages/cli/src/commands/database/seed/tenant.ts
@@ -1,11 +1,5 @@
 import { createTenantMetadata } from '@logto/core-kit';
-import type {
-  CreateTenant,
-  AdminData,
-  UpdateAdminData,
-  CreateScope,
-  CreateRolesScope,
-} from '@logto/schemas';
+import type { AdminData, UpdateAdminData, CreateScope, CreateRolesScope } from '@logto/schemas';
 import { generateStandardId } from '@logto/shared';
 import { assert } from '@silverhand/essentials';
 import type { CommonQueryMethods } from 'slonik';
@@ -18,7 +12,11 @@ import { getDatabaseName } from '../../../queries/database.js';
 export const createTenant = async (pool: CommonQueryMethods, tenantId: string) => {
   const database = await getDatabaseName(pool, true);
   const { parentRole, role, password } = createTenantMetadata(database, tenantId);
-  const createTenant: CreateTenant = { id: tenantId, dbUser: role, dbUserPassword: password };
+  const createTenant = {
+    id: tenantId,
+    dbUser: role,
+    dbUserPassword: password,
+  };
 
   await pool.query(insertInto(createTenant, 'tenants'));
   await pool.query(sql`
