+++ packages/cloud/package.json
@@ -31,7 +31,7 @@
     "@logto/shared": "workspace:^2.0.0",
     "@silverhand/essentials": "^2.5.0",
     "@withtyped/postgres": "^0.11.0",
-    "@withtyped/server": "^0.11.0",
+    "@withtyped/server": "^0.11.1",
     "accepts": "^1.3.8",
     "chalk": "^5.0.0",
     "decamelize": "^6.0.0",
