+++ packages/console/src/contexts/TenantsProvider.tsx
@@ -1,4 +1,5 @@
-import { type TenantInfo, TenantTag, defaultManagementApi } from '@logto/schemas';
+import { defaultManagementApi } from '@logto/schemas';
+import { type TenantInfo, TenantTag } from '@logto/schemas/models';
 import { conditional, noop } from '@silverhand/essentials';
 import type { ReactNode } from 'react';
 import { useCallback, useMemo, createContext, useState } from 'react';
