+++ packages/schemas/src/models/tenants.ts
@@ -1,7 +1,12 @@
 import { createModel } from '@withtyped/server/model';
+import type { InferModelType } from '@withtyped/server/model';
 import { z } from 'zod';
 
-import { TenantTag } from '../types/tenant/tag.js';
+export enum TenantTag {
+  Development = 'development',
+  Staging = 'staging',
+  Production = 'production',
+}
 
 export const Tenants = createModel(/* sql */ `
   /* init_order = 0 */
@@ -20,3 +25,11 @@ export const Tenants = createModel(/* sql */ `
 `)
   .extend('tag', z.nativeEnum(TenantTag))
   .extend('createdAt', { readonly: true });
+
+export type TenantModel = InferModelType<typeof Tenants>;
+
+export type TenantInfo = Pick<TenantModel, 'id' | 'name' | 'tag'> & { indicator: string };
+
+export const tenantInfoGuard: z.ZodType<TenantInfo> = Tenants.guard('model')
+  .pick({ id: true, name: true, tag: true })
+  .extend({ indicator: z.string() });
