+++ packages/schemas/src/seeds/tenant.ts
@@ -1,8 +1,4 @@
-import type { InferModelType } from '@withtyped/server/model';
-
-import type { Tenants } from '../models/tenants.js';
+export type { TenantModel } from '../models/tenants.js';
 
 export const defaultTenantId = 'default';
 export const adminTenantId = 'admin';
-
-export type TenantModel = InferModelType<typeof Tenants>;
