+++ packages/schemas/src/types/index.ts
@@ -18,4 +18,3 @@ export * from './theme.js';
 export * from './cookie.js';
 export * from './dashboard.js';
 export * from './domain.js';
-export * from './tenant/index.js';
