--- packages/schemas/src/types/tenant/tag.ts
@@ -1,5 +1,0 @@
-export enum TenantTag {
-  Development = 'development',
-  Staging = 'staging',
-  Production = 'production',
-}
